requirements.txt
flake8==4.0.1
flask==2.1.3
python-dotenv==0.20.0